Split SymbolExtractor Dockerfile helpers
